determine-test-runners: allow usage on macOS without `--dependents`
